refactoe part 2
