+++ cogs/autorespond.py
@@ -1,12 +1,7 @@
 from discord.ext import commands
-import discord
-import re
-import random
-import requests
-import asyncio
-import os
+from keys import EPIC_EMOJIS_LINK
+import discord, re, random, requests, asyncio
 
-EPIC_EMOJIS_LINK = os.environ['EPIC_EMOJIS_LINK']
 
 seals = """What the fuck did you just fucking say about me, you little bitch? I'll have you know I graduated top of 
 my class in the Navy Seals, and I've been involved in numerous secret raids on Al-Quaeda, and I have over 300
